Add instructions for installing Python and PyGame from their respective websites.
Mac OS X might come with Python, but it's always an old version. It's very
hard to get PyGame to run with Apple's Python, so we need to install a
newer version of Python from python.org. The advantage is that this also
comes bundled with PIP nowadays, so no need to use Homebrew anymore?

Installing PyGame on Mac OS X is slightly more cumbersome than on
Linux/Windows, since the installer is not properly signed. Some hoops
are necessary to make it work. Also, it only installs 32-bit libraries,
so one needs to run "python-32" in order for "import pygame" to work.

On Windows it is just a question of running two installers.